+++ Gemfile
@@ -1,20 +1,21 @@
 source 'https://rubygems.org'
 ruby '2.4.0'
 
-#group :test do
+gem 'rake'
+gem 'bcrypt'
+gem 'sinatra-flash'
+gem 'data_mapper', '1.2.0'
+gem 'dm-postgres-adapter'
+gem 'sinatra'
+gem 'sinatra-partial'
+
+group :test do
   gem 'cucumber'
   gem 'rspec'
-  gem 'rake'
   gem 'rubocop'
   gem 'simplecov', require: false
   gem 'simplecov-console', require: false
-  gem 'bcrypt'
-  gem 'sinatra-flash'
-  gem 'data_mapper', '1.2.0'
-  gem 'dm-postgres-adapter'
   gem 'capybara'
-  gem 'sinatra'
   gem 'database_cleaner'
   gem 'orderly'
-
-#end
+end
