+++ app/app.rb
@@ -1,90 +1,9 @@
 ENV['RACK_ENV'] ||= 'development'
 require 'sinatra/base'
 require 'sinatra/flash'
-require_relative 'models/peep'
+require 'sinatra/partial'
 require_relative 'data_mapper_setup'
-
-class Chitter < Sinatra::Base
-  enable :sessions
-  set :session_secret, 'session'
-  register Sinatra::Flash
-  include BCrypt
-  use Rack::MethodOverride
-
-  helpers do
-    def current_user
-      @current_user ||= User.get(session[:user_id])
-    end
-  end
-
-  get '/' do
-    redirect '/peeps'
-  end
-
-  get '/peeps' do
-    @peeps = Peep.all
-    erb :'peeps/index'
-  end
-
-  get '/peeps/new' do
-    if current_user
-      p session[:user_id]
-      erb :'peeps/new'
-    else
-      flash.keep[:login] = 'You must be logged in to do this'
-      redirect '/sessions/new'
-    end
-  end
-
-  post '/peeps' do
-    current_user
-    Peep.create(content: params[:content], time: Time.now, name: @current_user.name, username: @current_user.username
-    )
-    redirect '/peeps'
-  end
-
-  get '/users/new' do
-    @user = User.new
-    erb :'users/new'
-  end
-
-  post '/users' do
-    @user = User.create(email: params[:email],
-                name: params[:name],
-                username: params[:username],
-                password: params[:password],
-                password_confirmation: params[:password_confirmation])
-
-    if @user.save
-      session[:user_id] = @user.id
-      redirect '/peeps'
-    else flash.now[:error] = @user.errors.values.join("<br>")
-      erb :'users/new'
-    end
-  end
-
-  get '/sessions/new' do
-    erb :'sessions/new'
-  end
-
-  post '/sessions' do
-    @user = User.authenticate(params[:email], params[:password])
-    if @user
-      session[:user_id] = @user.id
-      redirect '/peeps'
-    else
-      flash.now[:error] = 'Email or password is incorrect'
-      erb :'sessions/new'
-    end
-  end
-
-  delete '/sessions' do
-    session[:user_id] = nil
-    flash.keep[:logout] = 'You have logged out successfully'
-    redirect to '/peeps'
-  end
-
-
-  run! if app_file == $0
-
-end
+require_relative 'server'
+require_relative 'controllers/peeps'
+require_relative 'controllers/users'
+require_relative 'controllers/sessions'
