+++ app/views/_welcome.erb
@@ -1,0 +1,8 @@
+<% if current_user %>
+  Welcome,
+  <%= current_user.name %>
+  <form method='post' action='/sessions'>
+    <input type='hidden' name='_method' value='delete'>
+    <input type='submit' value='Sign out'>
+  </form>
+<% end %>
