+++ app/views/layout.erb
@@ -7,29 +7,8 @@
   </head>
 
   <body>
-    <% if current_user %>
-      Welcome,
-      <%= current_user.name %>
-      <form method='post' action='/sessions'>
-        <input type='hidden' name='_method' value='delete'>
-        <input type='submit' value='Sign out'>
-      </form>
-    <% end %>
-    <% if flash[:logout] %>
-      <div id='logout'>
-        <%= flash[:logout] %>
-      </div>
-    <% end %>
-    <% if flash[:login] %>
-      <div id='login'>
-        <%= flash[:login] %>
-      </div>
-    <% end %>
-    <% if flash[:error] %>
-      <div id='error'>
-        <%= flash[:error] %>
-      </div>
-    <% end %>
+    <%= partial :welcome %>
+    <%= partial :flash %>
     <%= yield %>
   </body>
 </html>
