+++ app/views/peeps/new.erb
@@ -1,4 +1,6 @@
-<h2> Write a peep! </h2>
+<h2>
+  Write a peep!
+</h2>
 
 <form action='/peeps' method='post'>
   <input type='text' name='content' placeholder='Whats on your mind'>
