+++ app/views/sessions/new.erb
@@ -1,10 +1,12 @@
-<h2> Sign In </h2>
+<h2>
+  Sign In
+</h2>
 
 <form action="/sessions" method="post">
   <label for="email">
     Email:
   </label>
-  <input type="email" name="email" required>
+  <input type="email" name="email" required="required">
   <label for="pasword">
     Password:
   </label>
